requirements.txt
-e .
pytest

aws-cdk.assets==1.62.0
aws-cdk.aws-apigateway==1.63.0
aws-cdk.aws-applicationautoscaling==1.62.0
aws-cdk.aws-autoscaling-common==1.62.0
aws-cdk.aws-cloudwatch==1.62.0
aws-cdk.aws-codeguruprofiler==1.62.0
aws-cdk.aws-dynamodb==1.63.0
aws-cdk.aws-ec2==1.62.0
aws-cdk.aws-efs==1.62.0
aws-cdk.aws-events==1.62.0
aws-cdk.aws-iam==1.62.0
aws-cdk.aws-kms==1.62.0
aws-cdk.aws-lambda==1.62.0
aws-cdk.aws-logs==1.62.0
aws-cdk.aws-s3==1.62.0
aws-cdk.aws-s3-assets==1.62.0
aws-cdk.aws-sns==1.62.0
aws-cdk.aws-sns-subscriptions==1.62.0
aws-cdk.aws-sqs==1.62.0
aws-cdk.aws-ssm==1.62.0
aws-cdk.cloud-assembly-schema==1.62.0
aws-cdk.core==1.62.0
aws-cdk.cx-api==1.62.0
aws-cdk.region-info==1.62.0
boto3==1.15.0
cdk-dynamo-table-viewer==3.1.2
